# crates/clawft-services/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeftOS gateway API",
    "version": "0.8.1",
    "description": "REST, WebSocket and MCP surface of `weft gateway`. Every route except GET /api/health needs a bearer token issued by the kernel daemon (`weft token issue`); there is no mint route. Request and response bodies are mostly untyped here; see docs/ui/api-reference.md. Hand-maintained: tests/openapi.rs fails when a served route is missing or stale."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/console/token": {
      "post": {
        "summary": "Mint a short-lived read-only token for one project from a tailnet identity",
        "operationId": "post_api_console_token",
        "tags": [
          "console"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "project"
                ],
                "properties": {
                  "project": {
                    "type": "string",
                    "description": "Project ULID"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{token, expires_at, project}"
          },
          "400": {
            "description": "Bad body"
          },
          "403": {
            "description": "Not a tailnet peer, login not allowlisted, or cross-origin"
          },
          "404": {
            "description": "Unknown project, or tailnet identity is not enabled"
          },
          "503": {
            "description": "whois or daemon unavailable"
          }
        },
        "description": "Disabled (404) unless gateway.tailnetIdentity.enabled. No bearer: the caller is identified by its TCP peer address and tailscale whois."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Daemon-issued owner token (wft_...). Project tokens are refused."
      }
    },
...
  }
}